third_party/WebKit/LayoutTests/external/wpt/magnetometer/idlharness.https-expected.txt

This is a testharness.js-based test.
PASS Test IDL implementation of Magnetometer Sensor
PASS Sensor interface: existence and properties of interface object
PASS Sensor interface object length
PASS Sensor interface object name
PASS Sensor interface: existence and properties of interface prototype object
PASS Sensor interface: existence and properties of interface prototype object's "constructor" property
PASS Sensor interface: existence and properties of interface prototype object's @@unscopables property
PASS Sensor interface: attribute activated
PASS Unscopable handled correctly for activated property on Sensor
PASS Sensor interface: attribute hasReading
PASS Unscopable handled correctly for hasReading property on Sensor
PASS Sensor interface: attribute timestamp
PASS Unscopable handled correctly for timestamp property on Sensor
PASS Sensor interface: operation start()
PASS Unscopable handled correctly for start() on Sensor
PASS Sensor interface: operation stop()
PASS Unscopable handled correctly for stop() on Sensor
PASS Sensor interface: attribute onreading
PASS Unscopable handled correctly for onreading property on Sensor
PASS Sensor interface: attribute onactivate
PASS Unscopable handled correctly for onactivate property on Sensor
PASS Sensor interface: attribute onerror
PASS Unscopable handled correctly for onerror property on Sensor
PASS Magnetometer interface: existence and properties of interface object
PASS Magnetometer interface object length
PASS Magnetometer interface object name
PASS Magnetometer interface: existence and properties of interface prototype object
PASS Magnetometer interface: existence and properties of interface prototype object's "constructor" property
PASS Magnetometer interface: existence and properties of interface prototype object's @@unscopables property
PASS Magnetometer interface: attribute x
PASS Unscopable handled correctly for x property on Magnetometer
PASS Magnetometer interface: attribute y
PASS Unscopable handled correctly for y property on Magnetometer
PASS Magnetometer interface: attribute z
PASS Unscopable handled correctly for z property on Magnetometer
PASS Magnetometer must be primary interface of new Magnetometer();
PASS Stringification of new Magnetometer();
PASS Magnetometer interface: new Magnetometer(); must inherit property "x" with the proper type
PASS Magnetometer interface: new Magnetometer(); must inherit property "y" with the proper type
PASS Magnetometer interface: new Magnetometer(); must inherit property "z" with the proper type
PASS Sensor interface: new Magnetometer(); must inherit property "activated" with the proper type
PASS Sensor interface: new Magnetometer(); must inherit property "hasReading" with the proper type
PASS Sensor interface: new Magnetometer(); must inherit property "timestamp" with the proper type
PASS Sensor interface: new Magnetometer(); must inherit property "start()" with the proper type
PASS Sensor interface: new Magnetometer(); must inherit property "stop()" with the proper type
PASS Sensor interface: new Magnetometer(); must inherit property "onreading" with the proper type
PASS Sensor interface: new Magnetometer(); must inherit property "onactivate" with the proper type
PASS Sensor interface: new Magnetometer(); must inherit property "onerror" with the proper type
FAIL UncalibratedMagnetometer interface: existence and properties of interface object assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
FAIL UncalibratedMagnetometer interface object length assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
FAIL UncalibratedMagnetometer interface object name assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
FAIL UncalibratedMagnetometer interface: existence and properties of interface prototype object assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
FAIL UncalibratedMagnetometer interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
FAIL UncalibratedMagnetometer interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
FAIL UncalibratedMagnetometer interface: attribute x assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
PASS Unscopable handled correctly for x property on UncalibratedMagnetometer
FAIL UncalibratedMagnetometer interface: attribute y assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
PASS Unscopable handled correctly for y property on UncalibratedMagnetometer
FAIL UncalibratedMagnetometer interface: attribute z assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
PASS Unscopable handled correctly for z property on UncalibratedMagnetometer
FAIL UncalibratedMagnetometer interface: attribute xBias assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
PASS Unscopable handled correctly for xBias property on UncalibratedMagnetometer
FAIL UncalibratedMagnetometer interface: attribute yBias assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
PASS Unscopable handled correctly for yBias property on UncalibratedMagnetometer
FAIL UncalibratedMagnetometer interface: attribute zBias assert_own_property: self does not have own property "UncalibratedMagnetometer" expected property "UncalibratedMagnetometer" missing
PASS Unscopable handled correctly for zBias property on UncalibratedMagnetometer
FAIL UncalibratedMagnetometer must be primary interface of new UncalibratedMagnetometer(); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL Stringification of new UncalibratedMagnetometer(); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "x" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "y" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "z" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "xBias" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "yBias" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL UncalibratedMagnetometer interface: new UncalibratedMagnetometer(); must inherit property "zBias" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL Sensor interface: new UncalibratedMagnetometer(); must inherit property "activated" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL Sensor interface: new UncalibratedMagnetometer(); must inherit property "hasReading" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL Sensor interface: new UncalibratedMagnetometer(); must inherit property "timestamp" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL Sensor interface: new UncalibratedMagnetometer(); must inherit property "start()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL Sensor interface: new UncalibratedMagnetometer(); must inherit property "stop()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL Sensor interface: new UncalibratedMagnetometer(); must inherit property "onreading" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL Sensor interface: new UncalibratedMagnetometer(); must inherit property "onactivate" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
FAIL Sensor interface: new UncalibratedMagnetometer(); must inherit property "onerror" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: UncalibratedMagnetometer is not defined"
Harness: the test ran to completion.
